Kept thinking about the graphics pipeline and how much work it would be. Well, got up, and bam, now it's here. The compute pipeline goes and calculates an image, then the graphics pipeline samples this image in the fragment shader, mapping the image 1 to 1 onto itself. This is nice because now the image layout during the compute stage is more easy to mess with. I think I still need to have a seperate renderpass for the compute stage to avoid using vkCmdClearColorImage (it restricts the image layout really harshly and the only way around it is memory barriers to change the layout).
Next steps:
Find a better way to sync operations (currently swapchain sync is isolated from compute, so was using a fence when could probably use semaphores for some better performance).
Move the renderpass's into Renderer classes.
Create a scene of assets with the Model class.
Add a second SSBO for lights.
Make a simple render system that calls the compute pipeline & send models (with just a basic color for now) down to the compute stage via a SSBO and do some raytracing.