I believe the multiple-rays-per-pixel issue may be partially fixed. It seems the main source was the random number generation method. It was only based on the randomState (changes once per frame) and the pixel coordinate, meaning mulitple rays from the same pixel would have the same random values because each additional ray per pixel is done by a vkCmdDispatch. This was solved by storing another value inside the alpha channel of the storage image. This is then used in the seeding of the random generator, allowing past dispatches to force diviation without having to flush the ubo with a new random state.
There does still appear to be an issue with light intensity. In my cpu raytracer, everything started to smooth out with higher rays per pixel, but things stay a bit rough and also get brighter as more rays are cast. will have to look into the cause.

this raytracer is doing the work of the cpu raytracer very quickly (~990x speed increase).